+++ BankexWallet/InterfaceElements/DesignElements/WalletSettings/SettingsDelegate.swift
